Update to a single-step build process with actual images.
Rather than having to go through a local volume to build an image with
an installed php, do it all in one step.  This makes quite a few changes
to the way the images are built.

The Dockerfiles are now identical and could be consolidated except that
the Docker Hub's automated build configuration only allows you to
specify the location of the Dockerfile and not the build context, so we
have to put the Dockerfile where we want the build to occur.